requirements.txt
numpy
pickle
pandas
streamlit
Pillow
scikit-learn